ENH: pandoc-fix-textile.py: add html hack back
In case of tables (and probably others) we still see html in the
pandoc-generated textile. Add the we-call-pandoc-to-convert-html hack
back. The key issue is to fiddle out only the html parts and convert
only those.

Add workaround for the edge case where the last line is also html, where
we would not print any converted text. This is b/c the end of the html
"block" is not detected since it is equal to EOF.